textproc/elasticsearch5: Use native facilites for manging the process
start/stop/status will behave correctly with procname set.

This now prevents Elastic from starting when a process is already
running, stops cleanly, and also handles multiple profiles correctly:
if only some of the profile processes are running, a "service
elasticsearch start" will only start the profiles that are not detected
as running already.

All of this comes with the caveat that Elastic is slow to start up and write
its pidfile, so if you are too quick it will allow duplicate processes
to start. Here be JVM dragons.

Approved by:	portmgr (with hat)
MFH:		2018Q1


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@462650 35697150-7ecd-e111-bb59-0022644237b5