New port: www/falkon, the successor to www/qupzilla-qt5
Qupzilla has migrated to the KDE umbrella upstream, and
renamed itself in the process. The first rebranded release
was 2018-02-27.

	https://davidrosca.blogspot.nl/2018/02/falkon-300-released.html

Qupzilla settings can be migrated to Falkon with
	cp -R ~/.config/qupzilla  ~/.config/falkon

In the medium term, qupzilla-qt5 will be deprecated and
users encouraged to move to falkon. qupzilla-qt4 is
based on an EOL toolkit and won't be touched.

Submitted by:	tcberner
Approved by:	tcberner (mentor, implicit)
Differential Revision:	https://reviews.freebsd.org/D13906


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@463644 35697150-7ecd-e111-bb59-0022644237b5