Add rubygem-commonmarker 0.17.8
CommonMarker is a Ruby wrapper for libcmark-gfm, GitHub's fork of the reference
parser for CommonMark. It passes all of the C tests, and is therefore
spec-complete. It also includes extensions to the CommonMark spec as documented
in the GitHub Flavored Markdown spec, such as support for tables,
strikethroughs, and autolinking.

WWW: https://github.com/gjtorikian/commonmarker


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@464278 35697150-7ecd-e111-bb59-0022644237b5