New port: graphics/waylandpp
Wayland is an object oriented display protocol, which features request
and events.  Requests can be seen as method calls on certain objects,
whereas events can be seen as signals of an object.  This makes the
Wayland protocol a perfect candidate for a C++ binding.

The goal of this library is to create such a C++ binding for Wayland
using the most modern C++ technology currently available, providing an
easy to use C++ API to Wayland.

WWW: https://github.com/NilsBrause/waylandpp

This port is ignored by default because it requires a build of
graphics/mesa-libs with WAYLAND=on which is off by default.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@465523 35697150-7ecd-e111-bb59-0022644237b5