net-p2p/bitcoin: Add an option to use newer BDB
Make the new option the default, so that when the port is recompiled it
will link against BDB 5.x or 6.x (whichever is the default in the system).
This will keep old wallet files readable and will not automatically
convert them in an incompatible with BDB 4.8 way. Newly created wallet
files will not be readable by Bitcoin Core versions linked with BDB 4.8.

Next thing to consider is to remove the option to link against BDB 4.8 as
this will remove the net-p2p/bitcoin -> databases/db48 dependency and will
make it possible to ditch the latter, which is scheduled for removal from
the ports collection on Apr 30 2018.

PR:		226527
Approved by:	kbowling (maintainer), antoine


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@465909 35697150-7ecd-e111-bb59-0022644237b5