sysutils/htcondor: Unbreak and upgrade to 8.4.12
Fixes build errors caused by increased strictness of C++ compilers and
the switch to ninja as default back-end for cmake

This is a conservative upgrade to maintain compatibility with the current
port. The latest release, 8.6.10, also builds on FreeBSD, but includes
major changes that need to be tested before being put into production.

Approved by:    jrm (mentor)
Differential Revision:  https://reviews.freebsd.org/D14909


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@466017 35697150-7ecd-e111-bb59-0022644237b5