audio/moony-lv2: Unbreak on 12
Due to bug#221341 cpow/clog functions are missing, and were implemented ad-hoc in the patch.
Recently clog has been added to 12 in r333577, which conflited with the ad-hoc implementation of clog in the patch.
To correct this, cmake-level cpow/clog detection were added, and ad-hoc implementations are now added
only when cpow/clog aren't present in the system.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@470362 35697150-7ecd-e111-bb59-0022644237b5