SC2091: Remove surrounding $() to avoid executing output.
ShellCheck has detected that you have a command that just consists of a
command substitution.

This is typically done in order to try to get the shell to execute a
command, because $(..) does indeed execute commands. However, it's also
replaced by the output of that command.

PR:		227109
Submitted by:	mat
Sponsored by:	Absolight


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@471270 35697150-7ecd-e111-bb59-0022644237b5