Recognize timezones on FreeBSD for the default install.
By default, FreeBSD copies the current timezone database to
/etc/localtime and stores the name of the symlink in /var/db/zoneinfo.
Qt expects /etc/localtime to be a symlink and uses the symlink's
target to infer the name of the default timezone.  The existing code
in Qt includes some workarounds for Linux distributions that also copy
the current timezone to /etc/localtime by checking for files that
contain the name of the timezone.  Add a similar check for
/var/db/zoneinfo on FreeBSD.

PR:		200763
Reviewed by:	tcberner
Differential Revision:	https://reviews.freebsd.org/D15792


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@472335 35697150-7ecd-e111-bb59-0022644237b5