Set OS versions to target system versions for CROSS_SYSROOT.
When cross-building a package via a CROSS_SYSROOT, set OSVERSION,
_OSRELEASE, and OSREL to the target OS version.  OSVERSION is derived
from the <sys/param.h> header in the sysroot.  _OSRELEASE cannot be
easily derived from a sysroot as it is a property of a kernel, so
instead generate a "fake" _OSRELEASE from the OSVERSION, but use a
"-CROSS" branch name that will hopefully fail in any places that make
assumptions about branch names.

Approved by:	portmgr (bapt)
Differential Revision:	https://reviews.freebsd.org/D15812


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@472890 35697150-7ecd-e111-bb59-0022644237b5