Enable .init_array and .fini_array with xtoolchain GCC.
For base/gcc, only enable .init_array and .fini_array for 12.0 and later
to match the policy used by clang and llvm.  For the xtoolchain ports,
enable these unconditionally since the only OS version handy is the OS
version of the build host, not the target OS to be built.  Unlike llvm,
the use of .init_array and .fini_array is set at compile-time for GCC
and can't be changed at runtime.

Bump PORTREVISION for base/gcc and the xtoolchain GCC ports.

Reviewed by:	bapt (earlier version)
Differential Revision:	https://reviews.freebsd.org/D16014


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@474469 35697150-7ecd-e111-bb59-0022644237b5