Properly set CONFIGURE_TARGET for when amd64 is either the host or target.
CONFIGURE_TARGET is used for --build which is the host performing the
build.  TARGETARCH is used in powerpc64-gcc/Makefile as the target
architecture of the binaries generated by the compiler being built.
When building a non-amd64 target compiler on an amd64 host,
CONFIGURE_TARGET was set to amd64-unknown-freebsd which confused GCC
(it breaks compilation of an i386 target compiler on an amd64 host).
It would also result in setting CONFIGURE_TARGET improperly when
building an amd64 compiler on a non-amd64 host.  Fix by remapping
amd64 to x86-64 if the ARCH is amd64 rather than the TARGETARCH.

Reviewed by:	bapt
Differential Revision:	https://reviews.freebsd.org/D16227


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@475291 35697150-7ecd-e111-bb59-0022644237b5