New port: games/adonthell-wastesedge
Waste's Edge game for the Adonthell engine

As a loyal servant of the elven Lady Silverhair, you arrive at the
remote trading post of Waste's Edge, where she is engaged in
negotiations with the dwarish merchant Bjarn Fingolson.  But not
all is well at Waste's Edge, and soon you are confronted with
circumstances that are about to destroy your mistress' high reputation.
And you are the only one to avert this ...

WWW: http://adonthell.nongnu.org/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@477591 35697150-7ecd-e111-bb59-0022644237b5