audio/xmms-sapplug: add -fPIC to linking with lld on i386 and arm64
By default lld disallows relocations against read-only segments and
requires that shared objects are built with -fPIC (as they should be).
This port already set -fPIC for amd64, so apply it to arm64 (which
already uses lld as the system linker) and i386 (which will soon switch
to lld as the system linker).

Approved by:	portmgr (lld blanket)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@480738 35697150-7ecd-e111-bb59-0022644237b5