japanese/spamassassin: Mark broken
Mark japanese/spamassassin since it does not build after the update of
mail/spamassassin to 3.4.2.  The patch for japanese support does not apply.

Reset maintainer mail address, since mail to it bounces (domain not found).

Deprecate and set expiration date to the end of the year.

Discussed with:	portmgr (rene, @ EuroBSDCon 2018)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@480764 35697150-7ecd-e111-bb59-0022644237b5