Sort various parts of this Makefile properly by pulling in ONLY_FOR_ARCHS earlier (and sorting its contents) as well as moving USES earlier and some unrelated variables later. [1]
Also remove an obsolete conflict with Grub 0.x.

Finally, though, mark this port for deprecation given that it's been
without a maintainer for five-and-a-half years, not updated to Grub 2.02
which was released one-and-a-half years ago, and fails to build with
modern compilers such as clang and GCC 8. [2]

PR:		232961 [2]
Reported by:	portlint [1]


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@484145 35697150-7ecd-e111-bb59-0022644237b5