sysutils/iocage-devel: development version replacing library
iocage is a jail/container manager amalgamating some of the best features and
technologies the FreeBSD operating system has to offer. It is geared for ease
of use with a simple and easy to understand command syntax. This development
version aims to replace the library with a more solid code base while keeping
the same CLI.

WWW: https://github.com/iocage/iocage


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@484319 35697150-7ecd-e111-bb59-0022644237b5