devel/spread-sheet-widget: Library for Gtk+ which provides a widget for tabular data
GNU Spread Sheet Widget is a library for Gtk+ which provides a widget
for viewing and manipulating 2 dimensional tabular data in a manner
similar to many popular spread sheet programs.

The design follows the model-view-controller paradigm and is of
complexity O(1) in both time and space. This means that it is efficient
and fast even for very large data.

Features commonly found in graphical user interfaces such as cut and
paste, drag and drop and row/column labelling are also included.

WWW: https://www.gnu.org/software/ssw/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@484679 35697150-7ecd-e111-bb59-0022644237b5