Add py-vincenty 0.1.4
Vincenty calculates the geographical distance (in kilometers or miles) between 2
points with extreme accuracy.

This library implements Vincenty's solution to the inverse geodetic problem. It
is based on the WGS 84 reference ellipsoid and is accurate to within 1 mm (!) or
better.

This formula is widely used in geographic information systems (GIS) and is much
more accurate than methods for computing the great-circle distance (which assume
a spherical Earth).

WWW: https://github.com/maurycyp/vincenty


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@486341 35697150-7ecd-e111-bb59-0022644237b5