Fix build when devel/rapidjson is installed
devel/leatherman include an old version of RapidJSON that is not compatible
with what devel/rapidjson install.  Reorder includes so that the version
included with devel/leatherman is found before the one of devel/rapidjson
because it is what is wanted.

While here, fix `make test`.

No need to bump PORTREVISION.

PR:		233426
Reported by:	vladislav.movchan@gmail.com


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@490027 35697150-7ecd-e111-bb59-0022644237b5