Add py-typing-extensions 3.7.2
The typing module was added to the standard library in Python 3.5 on a
provisional basis and will no longer be provisional in Python 3.7. However, this
means users of Python 3.5 - 3.6 who are unable to upgrade will not be able to
take advantage of new types added to the typing module, such as typing.Text or
typing.Coroutine.

The typing_extensions module contains both backports of these changes as well as
experimental types that will eventually be added to the typing module, such as
Protocol.

WWW: https://github.com/python/typing/tree/master/typing_extensions


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@490583 35697150-7ecd-e111-bb59-0022644237b5