x11-fonts/spleen: Use x11-fonts/vtfontcvt-ng
and drop other workarounds for the buggy/incomplete base vtfontcvt(8).

We still check for /usr/bin/vtfontcvt in the port as an easy way
to exclude the VT option on systems where users have opted out of
vt(4) entirely or on DragonFly where it does not make sense to
provide vt(4) fonts.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@490696 35697150-7ecd-e111-bb59-0022644237b5