New port: security/caesarcipher
caesarcipher is an implementation of the traditional Caesar cipher
encryption algorithm with an extension to all printable ASCII
characters.  It can be run as an interactive tool on the command
line or process input- and outputfiles that are passed to the
ultility.

WWW: http://www.olivermahmoudi.com/programming/caesarcipher

PR:		222760
Submitted by:	Oliver Mahmoudi <fbsd@olivermahmoudi.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@491137 35697150-7ecd-e111-bb59-0022644237b5