lang/php72 and lang/php73: improve documentation of php-fpm specific www.conf
If you only define a port to listen in www.conf, PHP defaults to listen
only to IPv6 ports on FreeBSD. On other OS it listens to IPv6 and IPv4.

Since upstream do not want to fix this [1], we add this special case to
the documentation.

Because the configuration file is a @sample it wont be updated for already
changed files. Therefore i do not bump PORTREVISION.

PR:		235141
Submitted by:	 Artyom Davidov <ard_1@mail.ru>

[1] https://bugs.php.net/bug.php?id=74166


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@491300 35697150-7ecd-e111-bb59-0022644237b5