graphics/sane-backends: Actually update to 1.0.27
The current version is 1.0.25 with a wrongly named distfile which
fails to fetch from the fossies.org mirror with a size mismatch.
The update to 1.0.27 in r444767 also shows that the checksum and
size have not changed in distinfo in that commit.  Full changes
between the distfiles from 1.0.27_4 & 1.0.27_5 can be found at [1].

- Remove dead alioth.debian.org mirror
- Remove some no longer needed patches

[1] https://people.freebsd.org/~tobik/logs/sane-backends-1.0.27.diff

ABI:		https://abi-laboratory.pro/tracker/timeline/sane-backends/
PR:		229653
Submitted by:	Peter Wullinger <freebsd@dhke.de>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@491557 35697150-7ecd-e111-bb59-0022644237b5