New port: sysutils/sandboxfs
Sandboxfs is a FUSE file system that exposes a combination of
multiple files and directories from the host's file system in the
form of a virtual tree with an arbitrary layout.  You can think of
a sandbox as an arbitrary view into the host's file system with
different access privileges per directory.

Sandboxfs is designed to allow running commands with limited access
to the file system by using the virtual tree as their new root, and
to do so consistently across a variety of platforms.

WWW: https://github.com/bazelbuild/sandboxfs

Announcement:	https://julio.meroh.net/2019/02/hello-sandboxfs-0.1.0.html


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@492299 35697150-7ecd-e111-bb59-0022644237b5