New port: shells/ammonite
A modern replacement for the Bash system shell. Provides a systems shell in
the high-level Scala language, letting you seamlessly mix system operations
with real code without the hassle or the frustration of trying to write
complex code in Bash.

Ammonite lets you use the Scala language for scripting purposes: in a
Read-Evaluate-Print-Loop (REPL), as scripts, as a library to use
in existing projects, or as a standalone systems shell.

WWW: https://ammonite.io/

PR:		235759
Submitted by:	Jens Grassel <jan0sch@mykolab.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@493275 35697150-7ecd-e111-bb59-0022644237b5