graphics/feh: Update to 3.1.3
Update graphics/feh to 3.1.3
Add two new options, XINERAMA and CURL, default to on (this was the previous
default, with no option to turn off).
XINERAMA enables or disablex X11 Xienrama support
CURL enables or disables cURL support.
Change the EXIF option to have both a on and off, to match the pattern of
other options.

ChangeLog:
https://feh.finalrewind.org/archive/3.1.3/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@493316 35697150-7ecd-e111-bb59-0022644237b5