Deprecate devel/cpuflags
It's supposed to "return appropriate gcc flags to optimise compilation
for the current CPU", but it does not recognize modern CPUs.
MASTER_SITES is dead, no upstream is given in pkg-descr.

pkgsrc has a newer version 1.43 from 3 years ago, but for modern
compilers just using -march=native is probably a better alternative.

PR:		235993


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@493762 35697150-7ecd-e111-bb59-0022644237b5