Mark broken on FreeBSD 12+
On FreeBSD 12 and FreeBSD 13, puppetserver5 fails to start.  The issue will
likely not be fixed in a timely fashion, so make it clear to end-users that
the port is broken.

As a workaround, users of FreeBSD 12.0 can install puppetserver5 in a
FreeBSD 11.2 jail.

More details, progress and ways to help can be found in the PR 233766.

These changes are linked to the PuppetServer 6 ones committed previously in
r495804.

With hat:	puppet
PR:		233766
Reported by:	mikael.urankar@gmail.com


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@495904 35697150-7ecd-e111-bb59-0022644237b5