java/bouncycastle15: install signed bcprov.jar
Previous revision 497549 switched to install bcprov.jar build from sources
instead of bundled version. However, the bundled version of JAR file is signed
that is essential as is provides an implementation of a Provider
in the Java Cryptography Architecture. For details, refer to
https://docs.oracle.com/javase/8/docs/technotes/guides/security/crypto/HowToImplAProvider.html

Partially back out that revision to install bundled bcprov.jar as before.

Thanks to Alex Dupre for reporting this.

Reported by:	ale


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@497741 35697150-7ecd-e111-bb59-0022644237b5