Fix version going backward due to update to 2020.0.0.a1 (pointy hat to yours truly).
The maintainer no longer wants to maintain ksh93 as his desire is to
maintain a backlevel port of ksh93 in order to build and support a
shared library for use by legacy applications, which is inconsistent
with the direction of the att/ast team on github. I will maintain the
port.

PR:		237332
PR:		238266
Approved by:	maintiner (saper@saper.info)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@503249 35697150-7ecd-e111-bb59-0022644237b5