devel/gn: Update to 1592
Changelog:

Minor fixes to tracing

The chrome://tracing expects the timestamp to be in microseconds, but
it had been stored in the TraceItem in nanoseconds, so this fixes it.

Also adds quotes around the thread_id, since on Mac it is written in hex
format and so the JSON parser doesn't recognize it.

Finally, adds toolchain information to the import traces.

PR:		238241
Submitted by:	Oleh Hushchenkov <o.hushchenkov@gmail.com> (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@503332 35697150-7ecd-e111-bb59-0022644237b5