Prepare for powerpc-on-clang by deleting hard-coded tests for architecture as a stand-in for "are we running on gcc".
In some cases we only need to specifically test for "are we on the ancient
base gcc", e.g, the usage of 'pragma'.

While here, in some cases turn off SSE functions more specifically based
on ARCH, and turn off -mtune=generic everywhere.  These are part of a
larger work in progress; these commits are for ports that would have
been touched by the the powerpc-on-clang test regardless.

Approved by:	portmgr (tier-2 blanket)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@507308 35697150-7ecd-e111-bb59-0022644237b5