security/fsh: Remove redundant escapes of ordinary characters
Makefile.in has a number of escapes (all following @ in the sed expression)
that only serve to stop a sed in the configure script from replacing them.
Escaping ordinary characters is considered undefined behavior by POSIX, and
the patch in bug #229925 aims to serve out errors for escaping of most
ordinary characters to reduce friction when some of these sequences are
given special meaning.

QA:
 * portlint: OK (minor warnings that should be addressed separately)
 * testport: OK (poudriere: 13.0-CURRENT and 12.0-RELEASE, amd64)
 * diffoscope: OK (only minor metadata changes to package)

PR:		240308
Approved by:	koobs (ports)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@511366 35697150-7ecd-e111-bb59-0022644237b5