Upgrade editors/openoffice-4 to 4.1.7.
The build wants to include the git hash value in the build artifact
for display on the Help -> About popup, but it can't determine that
when building from a source tarball, as opposed to a git checkout.
Patch around that until upstream implements a way to include this
info into the tarball and pick it up from there.

Add xorg to USES for correctness.

patch-i127929 has been committed upstream.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@512422 35697150-7ecd-e111-bb59-0022644237b5