x11/wdisplays: add new port
wdisplays is a graphical application for configuring displays in
Wayland compositors. It borrows some code from kanshi. It should work
in any compositor that implements the wlr-output-management-unstable-v1
protocol, including sway. The goal of this project is to allow precise
adjustment of display settings in kiosks, digital signage, and other
elaborate multi-monitor setups.

https://github.com/cyclopsian/wdisplays


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@513763 35697150-7ecd-e111-bb59-0022644237b5