sysutils/runit-faster: Remove default 'current' runlevel link on deinstall
'current' refers to the current runlevel and we cannot just delete it
during upgrades without disrupting or breaking systems that are
using a runlevel other than 'default'.  Removing it is out of the
question in the general case, but should be fine if we are using
'default' at the moment of the upgrade.  The 'default' link is
recreated again afterwards.

This will also silence Poudriere's leftovers warning for this port.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@514114 35697150-7ecd-e111-bb59-0022644237b5