devel/py-crc32c: create port
This package exposes to Python the crc32c algorithm implemented in the SSE 4.2
instruction set of Intel CPUs.

By default, if your CPU doesn't support this instruction, the package will fail
to load with an ImportError. If you still need to use the crc32c checksum
algorithm this package comes with a software implementation that can be loaded
instead.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@514275 35697150-7ecd-e111-bb59-0022644237b5