KDE Frameworks -- update to 5.63.0
KDE Frameworks are over 70 addon libraries to Qt which provide a wide variety
of commonly needed functionality in mature, peer reviewed and well tested
libraries with friendly licensing terms. For an introduction see the KDE
Frameworks web page.

This release is part of a series of planned monthly releases making
improvements available to developers in a quick and predictable manner.

Announcement:
	https://kde.org/announcements/kde-frameworks-5.63.0.php

PR:		241288
Exp-run by:	antoine


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@514669 35697150-7ecd-e111-bb59-0022644237b5