calc: Simplify and take maintainership
calc provides a horrific Makefile that uses ?= for nothing at all,
requiring the file to be mangled by hand. Still, it was a little
much for us to redefine every system command. `cp` is unlikely to
wander out of $PATH.

While here, add some options to not install some stuff. In particular,
make it possible to just install a simple usable calculator without
hundreds of scripts, functions, headers, and help files.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@516754 35697150-7ecd-e111-bb59-0022644237b5