Update KDE Plasma Desktop to 5.17.3
Tuesday, 12 November 2019. Today KDE releases a bugfix update to KDE Plasma 5,
versioned 5.17.3. Plasma 5.17 was released in October 2019 with many feature
refinements and new modules to complete the desktop experience.

This release adds a fortnight's worth of new translations and fixes from KDE's
contributors. The bugfixes are typically small but important and include:

 * Fix binding loop in lockscreen media controls. Commit. Fixes bug #413087.
 * [GTK3/Firefox] Fix scrollbar click region. Commit. Fixes bug #413118.
 * [effects/startupfeedback] Fallback to small icon size when no cursor size is configured.

Changelog:
	https://kde.org/announcements/plasma-5.17.2-5.17.3-changelog.php


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@517335 35697150-7ecd-e111-bb59-0022644237b5