net/tigervnc-{server,viewer}: Update to 1.10.1
This is a security release to fix a number of issues that were found by
Kaspersky Lab. These issues affect both the client and server and could
theoretically allow an malicious peer to take control over the software
on the other side.

No working exploit is known at this time, and the issues require the
peer to first be authenticated. We still urge users to upgrade when
possible.

Relnotes:	https://github.com/TigerVNC/tigervnc/releases/tag/v1.10.1
Sponsored by:	HAW International


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@520683 35697150-7ecd-e111-bb59-0022644237b5