Add rubygem-ms_rest 0.11.1
MsRestAzure is a library which supports the Azure clients (SDKs) generated with
Autorest tool. It contains core logic and helper classes for error handling and
authentication. Also it includes azure specific logic like long polling
functionality and Azure application authentication. Usually it is not supposed
to be used as a standalone gem but only as a dependency for generated client
gems.

WWW: https://github.com/Azure/azure-sdk-for-ruby
WWW: https://github.com/Azure/azure-sdk-for-ruby/tree/master/runtime/ms_rest_azure


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@521292 35697150-7ecd-e111-bb59-0022644237b5