vim: Add a MAKE_JOBS option for concurrent build, default off
Most systems are able to build vim in parallel most of the time.
The dependency chain in the Makefiles are a spaghetti mess and
concurrency is broken with some frequency, so it just isn't
feasible for it to be on by default.

It does shave quite a bit of time off the build (and I have had
.undef MAKE_JOBS_UNSAFE in Makefile.local for years), so now
there's an option. It's off by default.

PR:		243109
Reported by:	glewis


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@522175 35697150-7ecd-e111-bb59-0022644237b5