Update KDE Plasma Desktop to 5.18.1
Tuesday, 18 February 2020. Today KDE releases a feature update to KDE Plasma 5,
versioned 5.18.1. Plasma 5.18 was released in February 2020 with many feature
refinements and new modules to complete the desktop experience.

This release adds a week's worth of new translations and fixes from KDE's
contributors. The bugfixes are typically small but important and include:

Changelog:
	https://kde.org/announcements/plasma-5.18.0-5.18.1-changelog.php


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@526524 35697150-7ecd-e111-bb59-0022644237b5