Significantly simplify the logic to determine which port (or base version) of GCC to use based on the specification of USE_GCC.
This is based on the observation that we now only have a single version
of GCC in base, namely GCC 4.2, which is not in ports any longer.  And
we limit our choice to either the specific version requested or the
default version of GCC in the ports tree; i.e., we no longer consider
an installed port of any version in between (which is a fringe case
extremely few, if any, users would have experienced, and then only
outside a clean build environment in any case).

Streamline some debugging output accordingly.

Overall this removes some 25 lines of largely complex logic.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@526751 35697150-7ecd-e111-bb59-0022644237b5