x11/xdm: Fix generation of etc/X11/xdm/Xresources
By some stupidity of autotools, spacing for a sed regexp used to generate
Xresources gets broken, and the sed command doen't match as it should.  This
results in an Xresources file with an extra '#endif /* XPM */', which breaks
the file.
Patch Makefile.in to fix the regexp and have Xresources generated properly.

PR:		244404
Reported by:	olgeni
MFH:		2020Q1


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@527102 35697150-7ecd-e111-bb59-0022644237b5