Update pkg-message
As of MySQL 8.0.16, the MySQL server performs the upgrade tasks previously
handled by mysql_upgrade. Consequently, mysql_upgrade is unneeded and is
deprecated as of that version, and will be removed in a future MySQL version.
Because mysql_upgrade no longer performs upgrade tasks, it exits with status 0 unconditionally.

Sponsored by:	Netzkommune GmbH


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@527863 35697150-7ecd-e111-bb59-0022644237b5