Loosen version restrictions on runtime dependencies
It turns out that it is not so practical to require a very specific version
of a package using DISTVERSION because the dependency cannot be fulfilled
when the dependency port gets it PORTREVISION bumped.

While here, make it possible to run tests more than once in a row and set
standard testing environment. Also, fix a typo in RUN_DEPENDS.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@527906 35697150-7ecd-e111-bb59-0022644237b5