- Maintainer timeout after 4 months take Maintainership   Reported: 	2019-11-26 03:43 UTC by Kyle Evans   Modified: 	2020-03-11 21:02 UTC (History)
Total new update to CuraEngine
Cura saw a major rewrite between 15.04 and 4.3.0, hence the need for
PORTEPOCH -- the changeset is too large to enumerate. At a minimum,
the new version solves some issues that users will see attempting to
slice .stl files created from newer Blender versions.

A new port, libArcus, is also included -- this is a prereq for the new
version of Cura and will also be needed for the UI if it's eventually ported.

PR:		ports/242242


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@529154 35697150-7ecd-e111-bb59-0022644237b5