Add p5-Proc-Find-Parents 0.631
Return an array containing information about parent processes of $pid up to the
parent of all processes (usually "init"). If $pid is not mentioned, it defaults
to $$. The immediate parent is in the first element of array, followed by its
parent, and so on.

WWW: https://metacpan.org/release/Proc-Find-Parents

PR:		245445
Submitted by:	Zane C. Bowers-Hadley <vvelox@vvelox.net>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@531137 35697150-7ecd-e111-bb59-0022644237b5